# swagger.yaml
swagger: "2.0"
info:
  description: "API para restaurante/tienda de comidas. Los usuarios podran registrarse, loguearse, solicitar una lista de productos, hacer un pedido y consultar su pedido. Solo el usuario administrador podrá realizar las operaciones de edicion del estado de los pedidos y consulta de todos ellos. También este podrá crear, consultar, eliminar y actualizar productos."
  version: "1.0.0"
  title: "Delilah-Resto"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:3000"
basePath: "/v2"
tags:
- name: "usuarios"
  description: "Operaciones sobre los usuarios"
- name: "productos"
  description: "Operaciones sobre los productos"
- name: "pedidos"
  description: "Operaciones sobre los pedidos"
schemes:
- "https"
- "http"
paths:
  /crear_usuario:
    post:
      tags:
      - "usuarios"
      summary: "Crea un nuevo usuario"
      description: "Deberá especificar el role del usuario que se esta creando. Solo acepta 'Usuario' o 'Administrador'."
      operationId: ""
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Objeto usuario"
        required: true
        schema:
          $ref: "#/definitions/usuario"
      responses:
        "201":
          description: "Usuario agregado exitosamente"
          schema:
            $ref: "#/definitions/status_mensaje"
        "409":
          description: "El usuario ya existe"
          schema:
            $ref: "#/definitions/status_mensaje"
        "500":
          description: "Error inesperado"
          schema: 
            $ref: "#/definitions/status_mensaje"
  /login:
    post:
      tags:
      - "usuarios"
      summary: "Login with your username and password"
      description: "Responde con el token que necesitara para realizar el resto de las operaciones"
      operationId: ""
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: ""
        required: true
        schema:
          $ref: "#/definitions/usuario_login_req"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/usuario_login_res"
        "412":
          description: "Faltan datos"
          schema:
            $ref: "#/definitions/status_mensaje"
        "404":
          description: "El usuario ingresado no existe"
          schema:
            $ref: "#/definitions/status_mensaje"
        "403":
          description: "Contraseña incorrecta"
          schema:
            $ref: "#/definitions/status_mensaje"
        "500":
          description: "Error inesperado"
          schema: 
            $ref: "#/definitions/status_mensaje"
  /usuarios:
    get:
      security:
        - Bearer: []
      tags:
      - "usuarios"
      summary: "Obtiene la informacion del usuario"
      description: "Responde con la informacion del usuario logueado que solicita la accion. Tiene que enviar el token BEARER en el header de authorization"
      operationId: ""
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "header"
        name: "authorization"
        type: "string"
            
      responses:
        "200":
          description: "successful operation"
          schema:
              $ref: "#/definitions/usuario"
        "401":
          description: "UnauthorizedError"
          schema:
            $ref: "#/definitions/status_mensaje"
        "500":
          description: "Error inesperado"
          schema: 
            $ref: "#/definitions/status_mensaje"
  /productos:
    get:
      tags:
      - "productos"
      summary: "Obtiene todos los productos registrados"
      description: "Obtiene un array de los los productos."
      operationId: ""
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "header"
        name: "Authorization"
        description: "Tienes que estar logueado para realizar esta accion"
        required: false
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/array_productos"
        "401":
          description: "UnauthorizedError"
          schema:
            $ref: "#/definitions/status_mensaje"
        "500":
          description: "Error inesperado"
          schema: 
            $ref: "#/definitions/status_mensaje"
  /subir_producto:
    post:
      security:
        - Bearer: []
      tags:
      - "productos"
      summary: "Crea un nuevo producto"
      description: "Crea un nuevo producto."
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "header"
        name: "authorization"
        type: "string"
      - in: "body"
        name: "body"
        description: "Ingresa un objecto con las propiedades especificadas"
        required: true
        schema:
          $ref: "#/definitions/producto"
      responses:
        "200":
          description: "Producto agregado exitosamente"
          schema:
            $ref: "#/definitions/status_mensaje"
        "401":
          description: "UnauthorizedError"
          schema:
            $ref: "#/definitions/status_mensaje"
        "403":
          description: "Tienes que ser administrador para realizar esta accion"
          schema:
            $ref: "#/definitions/status_mensaje"
        "409":
          description: "El producto que quieres subir ya existe"
          schema:
            $ref: "#/definitions/status_mensaje"
        "500":
          description: "Error inesperado"
          schema: 
            $ref: "#/definitions/status_mensaje"
  /borrar_producto:
    delete:
      security:
        - Bearer: []
      tags:
      - "productos"
      summary: "Elimina un producto"
      description: "Elimina un producto"
      operationId: ""
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "header"
        name: "authorization"
        type: "string"
      - in: "body"
        name: "body"
        description: "Ingrese el nombre de producto que desea eliminar"
        required: true
        schema:
            $ref: "#/definitions/producto_get"
      responses:
        "200":
          description: "Producto eliminado exitosamente"
          schema:
            $ref: "#/definitions/status_mensaje"
        "401":
          description: "UnauthorizedError"
          schema:
            $ref: "#/definitions/status_mensaje"
        "403":
          description: "Tienes que ser administrador para realizar esta accion"
          schema:
            $ref: "#/definitions/status_mensaje"
        "404":
          description: "El producto que indicaste no existe"
          schema:
            $ref: "#/definitions/status_mensaje"
        "500":
          description: "Error inesperado"
          schema: 
            $ref: "#/definitions/status_mensaje"
  /actualizar_producto:
    put:
      security:
        - Bearer: []
      tags: 
      - "productos"
      summary: "Podrás modificar cualquier campo de un producto."
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "header"
        name: "authorization"
        type: "string"
      - in: "body"
        name: "body"
        description: ""
        required: true
        schema:
          $ref: "#/definitions/modificar_producto"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/status_mensaje"
        "401":
          description: "UnauthorizedError"
          schema:
            $ref: "#/definitions/status_mensaje"
        "403":
          description: "Tienes que ser administrador para realizar esta accion"
          schema:
            $ref: "#/definitions/status_mensaje"
        "404":
          description: "El producto que indicaste no existe"
          schema:
            $ref: "#/definitions/status_mensaje"
        "500":
          description: "Error inesperado"
          schema: 
            $ref: "#/definitions/status_mensaje" 
  /pedidos:
    get:
      security:
        - Bearer: []
      tags:
      - "pedidos"
      summary: "Obtiene la lista de pedidos"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "header"
        name: "authorization"
        type: "string"
      - in: "query"
        name: "id_pedido"
        type: "number"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/array_pedidos"
        "401":
          description: "UnauthorizedError"
          schema:
            $ref: "#/definitions/status_mensaje"
        "403":
          description: "Tienes que ser administrador para realizar esta accion"
          schema:
            $ref: "#/definitions/status_mensaje"
        "500":
          description: "Error inesperado"
          schema: 
            $ref: "#/definitions/status_mensaje"
  /my_order:
    get:
      security:
        - Bearer: []
      tags:
      - "pedidos"
      summary: "Obtiene la lista de los pedidos del usuario que hace el request"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "header"
        name: "authorization"
        type: "string"
      - in: "query"
        name: "id_pedido"
        type: "number"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/array_pedidos"
        "404":
          description: "successful operation"
          schema:
            $ref: "#/definitions/no_orders"
        "401":
          description: "UnauthorizedError"
          schema:
            $ref: "#/definitions/status_mensaje"
        "500":
          description: "Error inesperado"
          schema: 
            $ref: "#/definitions/status_mensaje"
  /nuevo_pedido:
    post:
      tags:
      - "pedidos"
      summary: "Crea un nuevo pedido"
      description: "Crea un nuevo pedido."
      operationId: ""
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "header"
        name: "authorization"
        type: "string"
      - in: "body"
        name: "body"
        description: "Crea un nuevo pedido"
        required: true
        schema:
          $ref: "#/definitions/pedido"
      responses:
        "200":
          description: "operacion exitosa"
          schema:
            $ref: "#/definitions/status_mensaje"
        "401":
          description: "UnauthorizedError"
          schema:
            $ref: "#/definitions/status_mensaje"
        "500":
          description: "Error inesperado"
          schema: 
            $ref: "#/definitions/status_mensaje"
  /update_state:
    put:
      tags:
      - "pedidos"
      summary: "Actualiza el estado de un pedido"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "header"
        name: "authorization"
        type: "string"
      - in: "body"
        name: "body"
        description: "Ingresa el numero de pedido y el nuevo estado"
        required: true
        schema:
          $ref: "#/definitions/pedido_update"
      responses:
        "200":
          description: "operacion exitosa"
          schema:
            $ref: "#/definitions/status_mensaje"
        "401":
          description: "UnauthorizedError"
          schema:
            $ref: "#/definitions/status_mensaje"
        "403":
          description: "Tienes que ser administrador para realizar esta accion"
          schema:
            $ref: "#/definitions/status_mensaje"
        "500":
          description: "Error inesperado"
          schema: 
            $ref: "#/definitions/status_mensaje"
  /delete_order:
    delete:
      tags:
      - "pedidos"
      summary: "Eliminar un pedido"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "header"
        name: "authorization"
        type: "string"
      - in: "body"
        name: "body"
        description: "Ingresa el numero de pedido"
        required: true
        schema:
          $ref: "#/definitions/delete_order"
      responses:
        "200":
          description: "operacion exitosa"
          schema:
            $ref: "#/definitions/status_mensaje"
        "401":
          description: "UnauthorizedError"
          schema:
            $ref: "#/definitions/status_mensaje"
        "403":
          description: "Tienes que ser administrador para realizar esta accion"
          schema:
            $ref: "#/definitions/status_mensaje"
        "500":
          description: "Error inesperado"
          schema: 
            $ref: "#/definitions/status_mensaje"
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header  
definitions:
  delete_order:
    type: "object"
    properties:
      id_pedido:
        type: "integer"
  no_orders:
    type: "object"
    properties:
      mensaje: 
        type: "string"
  producto_get:
      type: "object"
      properties:
        nombre: 
          type: "string"
  producto:
      type: "object"
      properties:
        id_producto: 
            type: "integer"
        nombre:
            type: "string"
        precio:
            type: "integer"
  array_productos:
      type: "array"
      items: 
        type: "object"
        properties:
          id_producto: 
            type: "integer"
          nombre:
            type: "string"
          precio:
            type: "integer"
  array_pedidos:
    type: "array"
    items:
      type: "object"
      properties:
        id_pedido:
          type: "integer"
        estado: 
          type: "string"
        usuario_id: 
          type: "integer"
        fecha: 
          type: "string"
        pago: 
          type: "string"
        direccion: 
          type: "string"
        precio_unitario: 
          type: "integer"
        id_producto: 
          type: "integer"
        cantidad: 
          type: "integer"
        nombre: 
          type: "string"
        precio: 
          type: "integer"
  modificar_producto:
      type: "object"
      properties:
        nombre:
          type: "string"
        campo:
          type: "string"
        nuevo_valor:
          type: "string"
  status_mensaje:
      type: "object"
      properties:
        status: 
          type: "string"
        mensaje:
          type: "string"
  pedido:
    type: "object"
    properties:
      id_pedido:
        type: "integer"
      estado: 
        type: "string"
      usuario_id: 
        type: "integer"
      pago: 
        type: "string"
      direccion: 
        type: "string"
      detalles:
        type: "array"
        items:
          type: "object"
          properties:
            id_producto:
              type: "integer"
            cantidad:
              type: "integer"
            precio_unitario:
              type: "integer"
  usuario_login_res:
    type: "object"
    properties:
      status:
        type: "string"
      usuario:
        type: "string"
      mensaje:
        type: "string"
      token:
        type: "string"
  usuario_login_req:
    type: "object"
    properties:
      usuario:
        type: "string"
      password:
        type: "string"
  usuario:
    type: "object"
    required:
    - "usuario_id"
    - "usuario"
    - "nombre_apellido"
    - "mail"
    - "telefono"
    - "direccion"
    - "contraseña"
    properties:
      usuario_id:
        type: "integer"
        format: "int64"
      usuario:
        type: "string"
        example: "nombredeusuario"
      nombre_apellido:
        type: "string"
      mail:
        type: "string"
      telefono:
        type: "integer"
      direccion:
        type: "string"
      password:
        type: "string"
      role:
        type: "string"
  pedido_update:
    type: "object"
    properties:
      id_pedido:
        type: "string"
      estado:
        type: "string"